added the whole thing'a'shabang'bang'a'bang'a'lang
the whole thing is pretty much like this, python environments are like backpacks full of sentient tools.

sometimes those tools get along, othertimes they try to kill eachother.

So normally for your main backpack you have to choose your tools carefully, and for expeditions where certain tools are a must, you create a 
virtual environment which is like a separate backpack with the same design, as the main backpack but different tools inside of it.

you following so far?

well, what if you want to create a whole army of backpacks, that each perform their own tasks, but communicate with a central manager backpack?
(pretend these are living backpacks like from a cartoon).

For each unique worker backpack to communicate with the manager backpack, they need to have all of the tools that the manager backpack has, so that the manager backpack can use all of its tools to interface with them, and tell them how to act in a shared tool language (they're alive but they communicate by pantomiming with their tools in hand and by mimicking or reacting to eachother's behaviors).

So every worker backpack needs to have the same contents as the manager backpack, and then optionally additional contents to perform additional duties.

this code is intended to metaphorically mass produce the metaphorical backpacks.

you set up the main virtual environment (the main backpack), then all of the worker backpacks exist inside of a root folder where they share all of their work with one another, then you create each worker one at a time, and each worker automatically is given the same resources as the manager backpack, each sub virtual environment has the same imports as the main virtual environment.

I, plan on using this to create a really complicated program with lots of libraries that might not normally get along in a single environment, but whom I need to have share work loads and data with eachother.

However my imagination is already running away at the thought of a python-only operating system, which is running in a minimalist OS, stripped of everything but a python compiler, and the whole OS code is a backpack so to speak with the main/universal functionalities of the OS existing as imported modules inside of the manager backpack, and each extension of the OS existing as a worker backpack, and each USER of the OS, existing as the same type of program ran as a sub-environment, so a worker backpack that is a manager to its own worker backpacks, and then when user customize their own desktop from raw command line code, to something useable that they like, they can modify the entire environment as its own backpack, and then create subprocesses that have their own environments within them for experimenting with newer or older versions of python or for making python programs which use different tool dependencies that are incompatible with what they have running their desktop.

It would be a creative person's heaven, and a productive person's hell. wouldn't that be an AWESOME educational tool?

we could call it "MakerOS" if the name isn't already taken. Hey, if you decide to take that name extend an invite to me, Niaschim I'd love to be the creative director or a lead designer or a resource researcher or something.... hmu if you undertake that task please, but you don't have to, to each their own, rock on everybody bye!!!!!!!!!!!!!!